Ensure premades won't crash

+++ app/src/main/java/fr/neamar/lolgamedata/pojo/Summoner.java
@@ -6,12 +6,12 @@ import org.json.JSONObject;
 import java.io.Serializable;
 
 public class Summoner implements Serializable {
-    public int id;
+    public long id;
     public String name;
     public int level;
 
     public Summoner(JSONObject summoner) throws JSONException {
-        id = summoner.getInt("id");
+        id = summoner.getLong("id");
         name = summoner.getString("name");
         level = summoner.getInt("level");
     }
